+++ App.xaml.cs
@@ -12,6 +12,9 @@ public partial class App : Application
 
 	protected override Window CreateWindow(IActivationState? activationState)
 	{
-		return new Window(new AppShell());
+		return new Window(new AppShell())
+		{
+			Title = AppVariant.PublicAppName
+		};
 	}
 }